Extract buildXAxisLayout into its own module
First safe step toward decomposing ChartContainer: move the pure
buildXAxisLayout helper (and its tick/time imports) into a dedicated
buildXAxisLayout.ts. Behavior-identical pure-function move; ChartContainer
drops ~110 lines (1453 -> 1250 across this branch).
